# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.createMaskRuleRequest:
    properties:
      custom_regex:
        type: string
      custom_template:
        type: string
      database:
        type: string
      datasource_id:
        type: integer
      field:
        type: string
      mask_type:
        type: string
      table_name:
        type: string
    type: object
  handler.createSensitiveTableRequest:
    properties:
      database:
        type: string
      datasource_id:
        type: integer
      sensitivity_level:
        type: string
      table_name:
        type: string
    type: object
...
  handler.updateAIConfigRequest:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      model:
        type: string
      provider:
        type: string
      timeout:
        type: string
    type: object
...
  handler.updateMaskRuleRequest:
    properties:
      custom_regex:
        type: string
      custom_template:
        type: string
      field:
        type: string
      mask_type:
        type: string
      table_name:
        type: string
    type: object
  handler.updateNotifyConfigRequest:
    properties:
      secret:
        type: string
      webhook_url:
        type: string
    type: object
...
  resp.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  resp.PageResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  resp.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: SQL 审核、查询和工单管理平台 API
  title: SQLFlow API
  version: "1.0"
paths:
...
  /mask-rules:
    get:
      description: 管理员获取脱敏规则列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 数据源ID
        in: query
        name: datasource_id
        type: string
      - description: 数据库名
        in: query
        name: database
        type: string
      - description: 表名
        in: query
        name: table_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.PageResponse'
        "500":
          description: 获取脱敏规则列表失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取脱敏规则列表
      tags:
      - 脱敏规则
    post:
      consumes:
      - application/json
      description: 管理员创建数据脱敏规则
      parameters:
      - description: 创建脱敏规则请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createMaskRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 创建脱敏规则失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建脱敏规则
      tags:
      - 脱敏规则
  /mask-rules/{id}:
    delete:
      description: 管理员删除指定脱敏规则
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的规则ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除脱敏规则
      tags:
      - 脱敏规则
    get:
      description: 管理员获取指定脱敏规则详情
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的规则ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取脱敏规则详情
      tags:
      - 脱敏规则
    put:
      consumes:
      - application/json
      description: 管理员更新指定脱敏规则
      parameters:
      - description: 规则ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新脱敏规则请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateMaskRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新脱敏规则
      tags:
      - 脱敏规则
...
  /sensitive-tables:
    get:
      description: 管理员获取敏感表列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 数据源ID
        in: query
        name: datasource_id
        type: string
      - description: 数据库名
        in: query
        name: database
        type: string
      - description: 表名
        in: query
        name: table_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.PageResponse'
        "500":
          description: 获取敏感表列表失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取敏感表列表
      tags:
      - 敏感表
    post:
      consumes:
      - application/json
      description: 管理员标记指定表为敏感表
      parameters:
      - description: 创建敏感表请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createSensitiveTableRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 创建敏感表失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建敏感表
      tags:
      - 敏感表
  /sensitive-tables/{id}:
    delete:
      description: 管理员删除指定敏感表标记
      parameters:
      - description: 记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的记录ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 记录不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除敏感表
      tags:
      - 敏感表
  /settings:
    get:
      description: 管理员获取钉钉通知和AI配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
      security:
      - BearerAuth: []
      summary: 获取系统设置
      tags:
      - 设置
  /settings/ai:
    put:
      consumes:
      - application/json
      description: 管理员更新AI审核服务的配置
      parameters:
      - description: AI配置请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateAIConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新AI配置
      tags:
      - 设置
  /settings/dingtalk:
    put:
      consumes:
      - application/json
      description: 管理员更新钉钉Webhook配置
      parameters:
      - description: 钉钉配置请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateNotifyConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新钉钉通知配置
      tags:
      - 设置
  /settings/dingtalk/test:
    post:
      description: 管理员发送一条测试钉钉通知消息
      produces:
      - application/json
      responses:
        "200":
          description: 测试消息已发送
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 钉钉通知未启用
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 测试钉钉通知
      tags:
      - 设置
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"